factor common utilities for scalar arguments to a new file
